refacotr: move MapperRegistry, UserMapper to slipp.mapper
